move billing periods back into the base package
