+++ support-models/src/main/scala/com/gu/support/redemptions/RedemptionData.scala
@@ -3,7 +3,6 @@ package com.gu.support.redemptions
 import cats.syntax.functor._
 import com.gu.support.encoding.Codec
 import com.gu.support.redemptions.redemptions.{CorporateAccountId, RedemptionCode}
-import com.gu.support.zuora.api.{ReaderType, Subscription}
 import io.circe.syntax._
 import io.circe.{Decoder, Encoder}
 
@@ -13,7 +12,6 @@ case class CorporateRedemption(override val redemptionCode: RedemptionCode, corp
 
 object RedemptionData {
     import Codec.deriveCodec
-    import com.gu.support.encoding.CustomCodecs.{decodeCountry, encodeCountryAsAlpha2}
     implicit val corporateCodec: Codec[CorporateRedemption] = deriveCodec
 
     implicit val encoder: Encoder[RedemptionData] = Encoder.instance {
