+++ support-models/src/main/scala/com/gu/support/workers/ProductTypeRatePlans.scala
@@ -1,8 +1,7 @@
 package com.gu.support.workers
 
-import com.gu.support.catalog.{Corporate, Product, ProductRatePlan}
+import com.gu.support.catalog.{Corporate, NoProductOptions, Product, ProductRatePlan}
 import com.gu.support.config.TouchPointEnvironment
-import com.gu.support.workers.BillingPeriod.{Quarterly, SixWeekly}
 
 
 trait ProductTypeRatePlans[P] {
