+++ support-models/src/test/scala/com/gu/support/workers/Fixtures.scala
@@ -2,7 +2,6 @@ package com.gu.support.workers
 
 import com.gu.i18n.Currency
 import com.gu.i18n.Currency.GBP
-import com.gu.support.workers.BillingPeriod.Monthly
 
 //noinspection TypeAnnotation
 object Fixtures {
