+++ support-services/src/test/scala/com/gu/support/promotions/ServicesFixtures.scala
@@ -4,7 +4,7 @@ import com.gu.support.catalog._
 import com.gu.support.config.TouchPointEnvironments
 import com.gu.support.config.TouchPointEnvironments.PROD
 import com.gu.support.promotions.DefaultPromotions.GuardianWeekly.NonGift
-import com.gu.support.workers.BillingPeriod.Annual
+import com.gu.support.workers.Annual
 import com.gu.support.zuora.api.{RatePlan, RatePlanData, Subscription, SubscriptionData}
 import org.joda.time.{DateTime, Days, LocalDate, Months}
