+++ support-workers/src/test/scala/com/gu/emailservices/SubscriptionEmailFieldHelpersSpec.scala
@@ -2,7 +2,6 @@ package com.gu.emailservices
 
 import com.gu.i18n.Currency.{EUR, GBP, USD}
 import com.gu.support.promotions.Promotions.SixForSixPromotion
-import com.gu.support.workers.BillingPeriod.{Annual, Monthly, Quarterly}
 import com.gu.support.workers._
 import org.joda.time.LocalDate
 import org.scalatest.flatspec.AnyFlatSpec
