+++ support-workers/src/test/scala/com/gu/monitoring/LambdaExecutionResultSpec.scala
@@ -4,8 +4,7 @@ import java.util.UUID
 
 import com.gu.i18n.Country
 import com.gu.i18n.Currency.GBP
-import com.gu.support.workers.BillingPeriod.Monthly
-import com.gu.support.workers.Contribution
+import com.gu.support.workers.{Contribution, Monthly}
 import io.circe.parser._
 import io.circe.syntax._
 import org.scalatest.flatspec.AnyFlatSpec
