+++ support-workers/src/test/scala/com/gu/support/workers/lambdas/CreatePaymentMethodStateDecoderSpec.scala
@@ -1,12 +1,13 @@
 package com.gu.support.workers.lambdas
 
 import com.gu.i18n.Currency.GBP
-import com.gu.support.workers.BillingPeriod.{Annual, Monthly}
+import com.gu.support.encoding.CustomCodecs._
 import com.gu.support.workers.JsonFixtures.{validBaid, _}
 import com.gu.support.workers._
 import com.gu.support.workers.states.CreatePaymentMethodState
 import com.typesafe.scalalogging.LazyLogging
 import io.circe.Json
+import io.circe.generic.auto._
 import io.circe.parser._
 import io.circe.syntax._
 import org.scalatest.flatspec.AnyFlatSpec
