+++ support-workers/src/test/scala/com/gu/zuora/ProductSubscriptionBuildersSpec.scala
@@ -7,8 +7,7 @@ import com.gu.i18n.Currency.GBP
 import com.gu.support.catalog.Domestic
 import com.gu.support.config.TouchPointEnvironments.SANDBOX
 import com.gu.support.promotions.PromotionService
-import com.gu.support.workers.BillingPeriod.Quarterly
-import com.gu.support.workers.GuardianWeekly
+import com.gu.support.workers.{GuardianWeekly, Quarterly}
 import com.gu.support.zuora.api.{Day, Month, ReaderType, SubscriptionData}
 import com.gu.zuora.ProductSubscriptionBuilders._
 import org.joda.time.LocalDate
